requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
cryptography==41.0.7
argon2-cffi==23.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0